feat: changed dotfiles to symlink setup
This commit changes the dotfiles to be stored in one central location
with symlinks to their usage locations, instead of using a repository
with the working tree set as the $HOME directory without symlinks.

This new approach works with git tools better; e.g. Magit doesn't handle
the previous approach very well