feat: latex support and improved text editing
Before this commit, non-code modes such as markdown and org would not
wrap lines neatly like most modern editors and would instead require
newlines to be inserted regularly into the text. This commit changes the
behaviour of emacs to wrap at the fill column and for line navigation to
navigate via screen lines rather than logical lines. This change is for
non-prog modes